corpus introspection (#1007)
* new run

* fixes

* change moved directory

* more schedule enhancement tests

* fix fore new weight variants

* add 300c release

* update config

* request experiment

* update to pre-300c commit id

* support bug benchmarks

* remove unnecessary imports

* more _optimal support

* more _optimal support

* new experiment

* fix compile on fuzzbench

* fix compile on fuzzbench

* fix bug in skim

* make presubmit happy

* fix yaml

* new targets have a larger map

* dummy

* aflplusplus_qemu fix

* presubmit fix

* remove temporary variants, update commit id

* fix compiles, update optimal to llvm12

* fix qemu

* remove temporary variants, update commit id

* request new experiment

* fix py

* fix lto

* request new experiment

* add afl++ and eclipser

* add afl++ and eclipser

* add afl++ and eclipser

* make presubmit happy

* fix presubmit

* fix a build for afl++ eclipser

* add double variant

* add double variant

* add double variant

* fix fuzzer.py

* fix presubmit

* fix presubmit

* introspection

* another introspection variant

* another introspection variant

* try a better schedule
