+++ .github/workflows/fuzzers.yml
@@ -56,7 +56,8 @@ jobs:
           - aflplusplus_cmplog
           - aflplusplus_dict2file
           - aflplusplus_eclipser
-          - aflplusplus_double
+          - aflplusplus_introspection2
+          - aflplusplus_schedule
 
         benchmark_type:
           - oss-fuzz
