+++ fuzzers/aflplusplus_introspection2/builder.Dockerfile
