+++ fuzzers/aflplusplus_optimal/fuzzer.py
@@ -57,8 +57,6 @@ def build():  # pylint: disable=too-many-branches,too-many-statements
         aflplusplus_fuzzer.build("tracepc", "dict2file")
     elif benchmark_name == 'php_php-fuzz-parser':
         aflplusplus_fuzzer.build("native", "cmplog", "dict2file")
-    elif benchmark_name == 'libhevc_hevc_dec_fuzzer':
-        aflplusplus_fuzzer.build("native", "cmplog", "dict2file")
     elif benchmark_name == 'proj4-2017-08-14':
         aflplusplus_fuzzer.build("tracepc", "cmplog")
     elif benchmark_name == 're2-2014-12-09':
