+++ fuzzers/aflplusplus_qemu/builder.Dockerfile
@@ -26,7 +26,7 @@ RUN cd / && wget https://github.com/ninja-build/ninja/releases/download/v1.10.1/
 
 # Download afl++
 RUN git clone https://github.com/AFLplusplus/AFLplusplus.git /afl && \
-    cd /afl && git checkout a4fd4ea0f46529feb09577a13cc7c053fb22146f
+    cd /afl && git checkout 688f4ffb89ebf41a497070e8fcf2927510b66874
     
 # Build afl++ without Python support as we don't need it.
 # Set AFL_NO_X86 to skip flaky tests.
