+++ fuzzers/aflplusplus_schedule/builder.Dockerfile
@@ -24,7 +24,7 @@ RUN apt-get update && \
 # Download and compile afl++.
 RUN git clone https://github.com/AFLplusplus/AFLplusplus.git /afl && \
     cd /afl && \
-    git checkout a4fd4ea0f46529feb09577a13cc7c053fb22146f
+    git checkout 0097b3d0baf49c47bcdcde0975865b75cef9a6ca
 
 # Build without Python support as we don't need it.
 # Set AFL_NO_X86 to skip flaky tests.
