+++ fuzzers/aflplusplus_schedule/description.md
@@ -1,13 +1,12 @@
-# aflplusplus + eclipser 2.0
+# aflplusplus
 
-AFL++ fuzzer instance that uses Eclipser 2.0
+AFL++ fuzzer instance that has the following config active for all benchmarks:
   - PCGUARD instrumentation 
-  - dict2file feature
+  - cmplog feature
   - "fast" power schedule
   - persistent mode + shared memory test cases
 
 Repository: [https://github.com/AFLplusplus/AFLplusplus/](https://github.com/AFLplusplus/AFLplusplus/)
-Repository: [https://github.com/SoftSec-KAIST/Eclipser](https://github.com/SoftSec-KAIST/Eclipser)
 
 [builder.Dockerfile](builder.Dockerfile)
 [fuzzer.py](fuzzer.py)
