+++ service/core-fuzzers.yaml
@@ -4,7 +4,6 @@ fuzzers:
     - afl
     - aflfast
     - aflplusplus
-    - aflplusplus_optimal
     - aflsmart
     - entropic
     - eclipser
